+++ test/clitest.hs
@@ -189,19 +189,19 @@ main = do
 
       -- both should fail with address 0x1234 -- a SPECIFIC address, not a ranomly generated one
       it "keccak-assumptions-setup" $ do
-        (exitCode, stdout, stderr) <- runForgeTest "test/contracts/fail/keccak-preimage-setup.sol" ["--num-cex-fuzz", "0"]
+        (exitCode, stdout, stderr) <- runForgeTest "test/contracts/fail/keccak-preimage-setup.sol" []
         stderr `shouldNotContain` "CallStack"
         stdout `shouldContain` "[validated]"
         stdout `shouldContain` "0x0000000000000000000000000000000000001234"
         exitCode `shouldBe` (ExitFailure 1)
       it "keccak-assumptions-constructor" $ do
-        (exitCode, stdout, stderr) <- runForgeTest "test/contracts/fail/keccak-preimage-constructor.sol" ["--num-cex-fuzz", "0"]
+        (exitCode, stdout, stderr) <- runForgeTest "test/contracts/fail/keccak-preimage-constructor.sol" []
         stderr `shouldNotContain` "CallStack"
         stdout `shouldContain` "[validated]"
         stdout `shouldContain` "0x0000000000000000000000000000000000001234"
         exitCode `shouldBe` (ExitFailure 1)
       it "keccak-wrong-preimage" $ do
-        (exitCode, stdout, stderr) <- runForgeTest "test/contracts/fail/keccak-wrong-preimage.sol" ["--num-cex-fuzz", "0"]
+        (exitCode, stdout, stderr) <- runForgeTest "test/contracts/fail/keccak-wrong-preimage.sol" []
         stderr `shouldNotContain` "CallStack"
         stdout `shouldContain` "[not reproducible]"
         exitCode `shouldBe` (ExitFailure 1)