+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## Added
+- Output geth compatible jsonl traces in `hevm exec` via `--json-trace`
 - Allow dumping unsolved SMT files via `--dump-unsolved`
 - Allow resolving unknown addresses to only those that are already deployed
   via option `--only-deployed`
@@ -38,6 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   threads of the symbolic execution
 - During running in --only-deployed mode, we forgot to force the address
   in the constraints to be the one we computed it to be. Fixed.
+- We now properly collect all storage reads from the program and build
+  a proper counterexample. Previously, some information might have been missing.
 
 ## Changed
 - Updated forge to 1.2.3 and forge-std to 60acb7aa (1.9.7+)
@@ -49,6 +52,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - hevm and solidity has been moved to under github.com/argotorg
 - Extra range constraints for ABI symbolic types are no longer added. They are
   not needed as they are enforced in the bytecode.
+- We no longer try to fuzz the Expr to find a concrete value that satisfies
+  the expression. This was not very effective and made the system more complex
+  to maintain. Echidna is an excellent fuzzer that can be used instead.
 
 ## [0.55.1] - 2025-07-22